Importer: Add random_id (#104)
* Importer: Add google_container_node_pool

This resource requires the name for import, but the name is optional. I've added the following:
1. Use of a GCP API client to fetch node pools for that cluster.
2. A function to ask the user which of the available node pools is the right one to import (this can't be guessed).
3. A function to ask for arbitrary input, if no choices.

The user input functions can be used in other resources too, where required fields are optional.

* Add to README

* Add --interactive flag

* Add tests for the user input functions

* Update flags in importer README.md

* Add test for fetchNodePools

* Add test for fetchNodePoolName

* Allow skipping import in interactive mode

* Importer: Add random_id

This is purely by asking the user to input it. It isn't stored anywhere except the state, so it can't be figured out.

* Fix typo

* Add interactive to upstream resources

* Address comments

* Add comment for fmt.Println

* Rearrange the fromUser functions

* Remove interactive with choices