[Breaking change] support cron cloud build jobs
This is also a breaking change which renames `branch_regex` to
`branch_name` in the `cicd` component and restrict the value to be the
exact name of the branch. The motivation behind this are:

- Enforce a 1:1 mapping between a branch to a GCP environment, which
means changes made to only one branch should trigger jobs associated
with one GCP environment (i.e. dev, staging or prod).

- Cloud Build job triggered by Cloud Scheduler requires exactly one
branch name to be specified.