add .terraform.lock.hcl to gitignore
Also adding it to our auto-generated .gitigore file because
* As we rely on CICD to run terraform and deploy the changes, this file
gets discarded during the deployment process.
* This file does not get updated automatically in each `terraform init`
run in order to lock the versions of provider.
People who are not familiar with this file and the deployment might
accidentially check in this file from their local `terraform init`
attempt in a directory and lock provider to bad or old versions. This
will cause CICD to not able to update this file and retrieve latest
provider versions.