Don't run both plan and apply for CSR
Cloud Source Repositories does not support code review or presubmit runs. If we set both plan and apply to run at the same time, they will conflict when trying to acquire the lock and may error out.

To get around this, set only auto-apply for non-prod envs, and only plan for the prod env.