Add passthrough to the project recipe
Also removed the error checking, it doesn't work here because one of the fields is optional and isn't always provided. It might be simpler this way and to just let the schema in the child recipe set certain fields as required or not.

Fixes https://github.com/GoogleCloudPlatform/healthcare-data-protection-suite/issues/951